requirements.txt
pygls >= v0.10
ruamel.yaml >= 0.16
